pyproject.toml
[build-system]
requires = ["setuptools>=67.4.0"]
build-backend = "setuptools.build_meta"

[project]
name = "brachinus"
version = "1.2.dev1"
authors = [
  { name="Juan Bindez", email="juanbindez780@gmail.com" },
]
description = "AES-256 CBC file encryption library with support for individual files and directory batch operations."
readme = "README.md"
requires-python = ">=3.7"
license = {text = "GPLv2 license"}
keywords = ["encryption", "crypto", "aes", "security", "files", "cryptography", "aes-256", "file-protection", "data-security",]
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"Environment :: Console",
	"Intended Audience :: Developers",
	"License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
	"Natural Language :: English",
	"Operating System :: OS Independent",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: 3.9",
	"Programming Language :: Python :: 3.10",
	"Programming Language :: Python :: 3.11",
	"Programming Language :: Python :: 3.12",
	"Programming Language :: Python",
	"Topic :: Security :: Cryptography",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: Terminals",
	"Topic :: System :: Filesystems",
	"Topic :: Utilities",
]
dependencies = ['pycryptodome']

[project.urls]
"Homepage" = "https://github.com/juanbindez/brachinus"
"Bug Reports" = "https://github.com/juanbindez/brachinus/issues"
"Read the Docs" = "http://brachinus.readthedocs.io/"

[project.scripts]
brachinus = "brachinus.cli:main"

[tool.setuptools.packages.find]
include = ["brachinus*"]

[tool.setuptools]
license-files = ["LICENSE"]